[confiscan] Handle symlinks correctly
Resolves #1

The for loop is replaced by a while-read loop fed by the output of
find, which lists all files regardless of whether a file or
directory has been passed as an argument.

This allows us to unanimously check for symlinks without having to
consider the file type (file or directory). If a working link is
found, it is followed and the original file is copied instead.
If a broken link is detected, it is skipped.